Change datatypes

+++ packages/contracts/src/systems/MapConfigSystem.sol
@@ -5,7 +5,7 @@ import { System } from "@latticexyz/world/src/System.sol";
 import { MapConfig } from "../codegen/index.sol";
 
 contract MapConfigSystem is System {
-  function createMapConfig(uint32 width, uint32 height, bytes32 seed) public {
+  function createMapConfig(uint8 width, uint8 height, bytes32 seed) public {
     MapConfig.set(width, height, seed);
   }
 }
